requirements.txt
torch==2.1.0
transformers==4.35.0
peft==0.6.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3